requirements.txt
Django==3.1.3
python-dateutil==2.8.1
djangorestframework==3.12.2
unittest-xml-reporting==3.0.4
requests==2.25.0